Refactor ui.js
