+++ app/js/state/utils.js
@@ -3,6 +3,16 @@
  * @module state/utils
  */
 
+// BN254 (alt_bn128) field modulus used in the ZK circuits
+export const BN254_MODULUS = BigInt('0x30644e72e131a029b85045b68181585d2833e84879b9709143e1f593f0000001');
+
+// Zero leaf for merkle trees - poseidon2("XLM"). Must match contract's get_zeroes()[0]
+export const ZERO_LEAF_HEX = '0x25302288db99350344974183ce310d63b53abb9ef0f8575753eed36e0118f9ce';
+
+// Tree depths - must match circuit and contract deployments
+export const TREE_DEPTH = 5;
+export const SMT_DEPTH = 5;
+
 /**
  * Converts hex string to Uint8Array.
  * @param {string} hex - Hex string (with or without 0x prefix)
@@ -100,4 +110,19 @@ export function normalizeU256ToHex(value) {
  */
 export function reverseBytes(bytes) {
     return bytes.reverse();
+}
+
+/**
+ * Converts hex string to LE bytes for Rust Merkle tree insertion.
+ * 
+ * The Rust Merkle tree uses from_le_bytes_mod_order (LE).
+ * Soroban stores U256 as BE and converts via BigUint::from_bytes_be.
+ * We reverse BE to LE to ensure the same numeric value.
+ * 
+ * @param {string} hex - Hex string (BE representation of U256)
+ * @returns {Uint8Array} LE bytes for Rust tree insertion
+ */
+export function hexToBytesForTree(hex) {
+    const beBytes = hexToBytes(hex);
+    return beBytes.reverse();
 }
