+++ app/js/ui/transactions/index.js
@@ -1,0 +1,9 @@
+/**
+ * Transaction modules index - re-exports all transaction modules.
+ * @module ui/transactions
+ */
+
+export { Deposit, setNotesTableRef as setDepositNotesTableRef } from './deposit.js';
+export { Withdraw, setNotesTableRef as setWithdrawNotesTableRef } from './withdraw.js';
+export { Transact, setNotesTableRef as setTransactNotesTableRef } from './transact.js';
+export { Transfer, setNotesTableRef as setTransferNotesTableRef } from './transfer.js';
